+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - MPMB: Corrently handle multi-classing
 - MPMB: Item importing: match if you can, or import the name into the Loot category
         - changed to batch Item update
+        - improved matching to find best match by intersecting words
 
 ## 0.6.0
 - Prototype importing from Fantasy Grounds